Refactor onboarding components into separate files
